+++ components/TextLoop/SimpleTextLoop.tsx
@@ -5,15 +5,21 @@ import { motion } from "framer-motion";
 interface TextLoopProps {
   loopItems: string[];
   delay: number;
+  className: string;
 }
 
 /**
  * Loops through an array of strings and displays them one by one.
  * @param loopItems (string[]) - Array of strings to loop through
  * @param delay (number) - Delay between each loop in milliseconds
+ * @param className (string) - Class name to apply to the span element
  * @returns (JSX.Element): A span element that loops through the given array of strings
  */
-const SimpleTextLoop: React.FC<TextLoopProps> = ({ loopItems, delay }) => {
+const SimpleTextLoop: React.FC<TextLoopProps> = ({
+  loopItems,
+  delay,
+  className,
+}) => {
   const [currentItemIndex, setCurrentItemIndex] = useState<number>(0);
 
   useEffect(() => {
@@ -32,14 +38,7 @@ const SimpleTextLoop: React.FC<TextLoopProps> = ({ loopItems, delay }) => {
       exit={{ opacity: 0, y: -20 }}
       transition={{ duration: 0.5 }}
       style={{ transitionProperty: "opacity, transform" }}
-      className="
-        text-2xl md:text-4xl 
-        font-semibold 
-        p-1 bg-clip-text text-transparent 
-        bg-gradient-to-r 
-        from-red-600 via-orange-500 to-rose-500 
-        dark:from-red-700 dark:via-orange-600 dark:to-rose-800
-        tracking-wide"
+      className={className}
     >
       {loopItems[currentItemIndex]}
     </motion.span>
