+++ components/TextLoop/TextLoop.tsx
@@ -1,10 +1,11 @@
 import React from "react";
-import TypewriterComponent from "typewriter-effect";
-import SimpleTextLoop from "../Atoms/SimpleTextLoop";
+import SimpleTextLoop from "./SimpleTextLoop";
+import TypeWriterTextLoop from "./TypeWriterTextLoop";
 
 type TextLoopProps = {
   loopItems: string[];
   implementation: "simple" | "typewriter";
+  className: string;
 };
 
 /**
@@ -15,34 +16,25 @@ type TextLoopProps = {
  * @param loopItems (string[]) - Array of strings to loop through
  * @param implementation ("simple" | "typewriter") - The implementation to use for looping through the text
  */
-const TextLoop: React.FC<TextLoopProps> = ({ loopItems, implementation }) => {
+const TextLoop: React.FC<TextLoopProps> = ({
+  loopItems,
+  implementation,
+  className,
+}) => {
   return (
     <>
       {/* Type Writer Implementation */}
       {implementation === "typewriter" && (
-        <p
-          className="
-            text-2xl md:text-4xl 
-            font-semibold 
-            p-1 bg-clip-text text-transparent 
-            bg-gradient-to-r 
-            from-red-600 via-orange-500 to-rose-500 
-            dark:from-red-700 dark:via-orange-600 dark:to-rose-800
-            tracking-wide"
-        >
-          <TypewriterComponent
-            options={{
-              strings: loopItems,
-              autoStart: true,
-              loop: true,
-            }}
-          />
-        </p>
+        <TypeWriterTextLoop className={className} loopItems={loopItems} />
       )}
 
       {/* Simple Loop Implementation */}
       {implementation === "simple" && (
-        <SimpleTextLoop loopItems={loopItems} delay={3000} />
+        <SimpleTextLoop
+          loopItems={loopItems}
+          delay={3000}
+          className={className}
+        />
       )}
     </>
   );
